+++ src/Test/QuickCheck/StateModel/Lockstep/Defaults.hs
@@ -27,8 +27,7 @@ import           Data.Typeable
 import qualified Test.QuickCheck as QC
 import           Test.QuickCheck (Gen, Property)
 import           Test.QuickCheck.StateModel (Action, Any (..), LookUp,
-                     PostconditionM, Realized, StateModel (Error), Var,
-                     monitorPost)
+                     PostconditionM, RunModel (Error), Var, monitorPost)
 import           Test.QuickCheck.StateModel.Variables (HasVariables (..),
                      VarContext)
 
@@ -99,8 +98,8 @@ postcondition :: forall m state a.
      RunLockstep state m
   => (Lockstep state, Lockstep state)
   -> LockstepAction state a
-  -> LookUp m
-  -> Realized m a
+  -> LookUp
+  -> a
   -> PostconditionM m Bool
 postcondition = postconditionWith True
 
@@ -114,8 +113,8 @@ postconditionWith :: forall m state a.
   => Bool -- ^ Verbose output
   -> (Lockstep state, Lockstep state)
   -> LockstepAction state a
-  -> LookUp m
-  -> Realized m a
+  -> LookUp
+  -> a
   -> PostconditionM m Bool
 postconditionWith verbose (before, _after) action _lookUp a =
     case checkResponse (Proxy @m) before action a of
@@ -129,8 +128,8 @@ monitoring :: forall m state a.
   => Proxy m
   -> (Lockstep state, Lockstep state)
   -> LockstepAction state a
-  -> LookUp m
-  -> Either (Error (Lockstep state)) (Realized m a)
+  -> LookUp
+  -> Either (Error (Lockstep state) m) a
   -> Property -> Property
 monitoring _p (before, after) action _lookUp _realResp =
       QC.counterexample ("State: " ++ show after)
@@ -175,7 +174,7 @@ instance InLockstep state => HasVariables (Action (Lockstep state) a) where
 checkResponse :: forall m state a.
      RunLockstep state m
   => Proxy m
-  -> Lockstep state -> LockstepAction state a -> Realized m a -> Either String String
+  -> Lockstep state -> LockstepAction state a -> a -> Either String String
 checkResponse p (Lockstep state env) action a =
     compareEquality
       (a         , observeReal p action a)
@@ -185,7 +184,7 @@ checkResponse p (Lockstep state env) action a =
     modelResp = fst $ modelNextState action env state
 
     compareEquality ::
-         (Realized m a, Observable state a)
+         (a, Observable state a)
       -> (ModelValue state a, Observable state a) -> Either String String
     compareEquality (realResp, obsRealResp) (mockResp, obsMockResp)
       | obsRealResp == obsMockResp = Right $ concat [