+++ src/Test/QuickCheck/StateModel/Lockstep/Op/SumProd.hs
@@ -1,8 +1,7 @@
 module Test.QuickCheck.StateModel.Lockstep.Op.SumProd (Op(..), intOpId) where
 
 import           Control.Monad ((<=<))
-import           Control.Monad.Reader (ReaderT)
-import           Control.Monad.State (StateT)
+import           Control.Monad.Identity (Identity)
 import           GHC.Show (appPrec)
 
 import           Test.QuickCheck.StateModel.Lockstep.Op
@@ -44,16 +43,8 @@ intOpId (OpComp g f) = intOpId g <=< intOpId f
 instance Operation Op where
   opIdentity = OpId
 
-instance InterpretOp Op (WrapRealized IO) where
-  intOp = intOpRealizedId intOpId
-
-instance InterpretOp Op (WrapRealized m)
-      => InterpretOp Op (WrapRealized (StateT s m)) where
-  intOp = intOpTransformer
-
-instance InterpretOp Op (WrapRealized m)
-      => InterpretOp Op (WrapRealized (ReaderT r m)) where
-  intOp = intOpTransformer
+instance InterpretOp Op Identity where
+  intOp = intOpIdentity intOpId
 
 {-------------------------------------------------------------------------------
   'Show' and 'Eq' instances