+++ src/Test/QuickCheck/StateModel/Lockstep/Run.hs
@@ -94,7 +94,7 @@ labelActions (Actions steps) =
 
 runActions ::
      ( RunLockstep state IO
-     , e ~ Error (Lockstep state)
+     , e ~ Error (Lockstep state) IO
      , forall a. IsPerformResult e a
      )
   => Proxy state
@@ -111,7 +111,7 @@ runActions _ actions = monadicIO $ void $ StateModel.runActions  actions
 -- is a reasonable choice.
 runActionsBracket ::
      ( RunLockstep state m
-     , e ~ Error (Lockstep state)
+     , e ~ Error (Lockstep state) m
      , forall a. IsPerformResult e a
      )
   => Proxy state